wip: refactor to own component
